# generated-code/printshop-v2/src/main/java/com/printshop/mis/domain/PaymentRecord.java
package com.printshop.mis.domain;

import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.persistence.Table;
import java.math.BigDecimal;

@Entity
@Table(name = "payments")
public class PaymentRecord {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    public Long id;
    public String paymentNo;
    public Long orderId;
    public BigDecimal amount = BigDecimal.ZERO;
    public String method;
    public String status;
    public String paidAt;
}