# obsidian-vault/wiki/summaries/API契约/OpenAPI-接口契约-v2.2.yaml
openapi: 3.0.3
info:
  title: 图文快印门店连锁管理系统 Print MIS API
  version: 2.2.0
  description: 准生产版 v2 接口契约，在 v2.1 变更冻结基础上追加统一工作流动作、角色任务工作台和订单聚合详情。
servers:
  - url: http://127.0.0.1:8080
security:
  - basicAuth: []
...
paths:
...
  /api/orders:
    get:
      tags: [Order]
      summary: 订单列表
      description: 按当前登录用户角色与门店过滤；客户仅返回本人订单，门店角色仅返回本门店订单，配送角色仅返回本人或待接单配送相关订单。
      responses:
        '200':
          description: 按当前角色过滤的订单列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [Order]
      summary: 创建订单
      description: customerId、customerName、dueAt、totalAmount 由服务端生成或计算；客户端只能提交订单规格、数量和交付选项。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrintOrder'
      responses:
        '200':
          description: 创建后的订单
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/orders/{id}:
    get:
      tags: [Order]
      summary: 订单详情
      description: 仅允许访问当前登录用户可见范围内的订单。
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: 订单详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    put:
      tags: [Order]
      summary: 更新订单
      description: 按角色限制可修改字段；totalAmount 由服务端重新计算，客户端传入值不会作为可信金额。
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrintOrder'
      responses:
        '200':
          description: 更新后的订单
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      tags: [Order]
      summary: 删除订单
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: 已删除订单
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/orders/{id}/status:
    post:
      tags: [Order]
      summary: 流转订单状态
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusChange'
      responses:
        '200':
          description: 更新后的订单
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/orders/{orderId}/aggregate:
    get:
      tags: [Order]
      summary: 订单聚合详情
      description: 返回订单、报价、作业单、生产、配送、财务、变更、文件和可执行下一步动作；客户和配送角色不返回审计时间线。
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 当前角色可见的订单聚合视图
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/orders/{id}/files:
    get:
      tags: [File]
      summary: 订单文件列表
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: 文件列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [File]
      summary: 上传订单文件
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: 文件元数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/order-change-requests/{id}:
    get:
      tags: [OrderChangeRequest]
      summary: 订单变更请求详情
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: 订单变更请求详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/orders/{orderId}/change-requests:
    post:
      tags: [OrderChangeRequest]
      summary: 创建订单变更请求
      description: 订单进入处理后，产品类型、颜色工艺、页数、份数、交付方式或加急信息变更必须创建请求，冻结生产/SLA，等待店长审批。
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderChangeRequestInput'
      responses:
        '200':
          description: 已创建的待审批变更请求
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/order-change-requests/{id}/approve:
    post:
      tags: [OrderChangeRequest]
      summary: 审批通过订单变更请求
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeDecision'
      responses:
        '200':
          description: 变更已批准，订单字段和金额已重新计算，生产/SLA 冻结解除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/order-change-requests/{id}/reject:
    post:
      tags: [OrderChangeRequest]
      summary: 驳回订单变更请求
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeDecision'
      responses:
        '200':
          description: 变更已驳回，订单保持原字段，生产/SLA 冻结解除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/job-tickets:
    get:
      tags: [JobTicket]
      summary: 作业单列表
      responses:
        '200':
          description: 作业单列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [JobTicket]
      summary: 创建作业单
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobTicket'
      responses:
        '200':
          description: 创建后的作业单
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/job-tickets/{id}:
    get:
      tags: [JobTicket]
      summary: 作业单详情
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: 作业单详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    put:
      tags: [JobTicket]
      summary: 更新作业单
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobTicket'
      responses:
        '200':
          description: 更新后的作业单
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      tags: [JobTicket]
      summary: 删除作业单
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: 已删除作业单
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /api/invoices:
    get:
      tags: [Finance]
      summary: 发票列表
      responses:
        '200':
          description: 发票列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [Finance]
      summary: 创建发票申请
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceRecord'
      responses:
        '200':
          description: 创建后的发票
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/invoices/{id}:
    get:
      tags: [Finance]
      summary: 发票详情
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: 发票详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    put:
      tags: [Finance]
      summary: 更新发票
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceRecord'
      responses:
        '200':
          description: 更新后的发票
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      tags: [Finance]
      summary: 删除发票
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: 已删除发票
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/invoices/{id}/issue:
    post:
      tags: [Finance]
      summary: 开具发票
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: 已开具发票
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/payments:
    get:
      tags: [Finance]
      summary: 收款/退款列表
      responses:
        '200':
          description: 付款记录列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags: [Finance]
      summary: 新增收款
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRecord'
      responses:
        '200':
          description: 创建后的付款记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/payments/{id}:
    get:
      tags: [Finance]
      summary: 付款详情
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: 付款详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    put:
      tags: [Finance]
      summary: 更新付款记录
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRecord'
      responses:
        '200':
          description: 更新后的付款记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      tags: [Finance]
      summary: 删除付款记录
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: 已删除付款记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/payments/{id}/refund:
    post:
      tags: [Finance]
      summary: 退款
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: 已退款付款记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          nullable: true
        message:
          type: string
        traceId:
          type: string
...
    StatusChange:
      type: object
      properties:
        status:
          type: string
        step:
          type: string
...
    PrintOrder:
      type: object
      properties:
        id: { type: integer, format: int64 }
        orderNo: { type: string }
        customerId: { type: integer, format: int64 }
        customerName: { type: string }
        storeId: { type: integer, format: int64 }
        storeName: { type: string }
        productType:
          type: string
          enum: [论文胶装, 培训手册, 名片快印, 海报写真, 宣传单页, 写真展板]
          description: 固定产品类型选项。
        colorMode:
          type: string
          enum: [黑白, 彩色, 黑白加彩页, 覆膜, 装订加覆膜]
          description: 固定颜色与后道工艺选项。
        pageCount: { type: integer }
        copies: { type: integer }
        dueAt:
          type: string
          description: 创建订单时默认等于服务端创建时间。
        deliveryMode:
          type: string
          enum: [到店自提, 同城配送, 跨店配送, 外协配送]
        priority:
          type: string
          enum: [普通, 加急, 特急]
        status: { type: string }
        paymentStatus: { type: string }
        currentStep: { type: string }
        totalAmount:
          type: number
          readOnly: true
          description: 服务端根据产品类型、颜色工艺、页数、份数、交付方式和优先级自动计算。
        paidAmount: { type: number }
...
    JobTicket:
      type: object
      properties:
        id: { type: integer, format: int64 }
        ticketNo: { type: string }
        orderId: { type: integer, format: int64 }
        quotationId: { type: integer, format: int64 }
        specs: { type: string }
        paperType: { type: string }
        binding: { type: string }
        status: { type: string }
...
    InvoiceRecord:
      type: object
      properties:
        id: { type: integer, format: int64 }
        invoiceNo: { type: string }
        orderId: { type: integer, format: int64 }
        title: { type: string }
        taxNo: { type: string }
        amount: { type: number }
        status: { type: string }
        issuedAt: { type: string }
    PaymentRecord:
      type: object
      properties:
        id: { type: integer, format: int64 }
        paymentNo: { type: string }
        orderId: { type: integer, format: int64 }
        amount: { type: number }
        method: { type: string }
        status: { type: string }
        paidAt: { type: string }
    OrderChangeRequestInput:
      type: object
      required: [reason]
      properties:
        productType:
          type: string
          enum: [论文胶装, 培训手册, 名片快印, 海报写真, 宣传单页, 写真展板]
        colorMode:
          type: string
          enum: [黑白, 彩色, 黑白加彩页, 覆膜, 装订加覆膜]
        pageCount:
          type: integer
          minimum: 1
        copies:
          type: integer
          minimum: 1
        deliveryMode:
          type: string
          enum: [到店自提, 同城配送, 跨店配送, 外协配送]
        priority:
          type: string
          enum: [普通, 加急, 特急]
        reason:
          type: string
          description: 变更理由。服务端记录原字段、新字段、金额差异、冻结开始时间和发起人。
    ChangeDecision:
      type: object
      properties:
        comment:
          type: string
          description: 店长或管理员审批意见。